Added options, transferred minimal strategy into rGS, rewrote tests
Options to select the coefficients and the variable name for the generated polynomial ring were added.

Code for generating the minimal generating sets was transferred into the main randomGeneratingSet function.

Tests on randomMinGeneratingSet were moved to tests on randomGeneratingSet with the minimal strategy option.

Tests for randomGeneratingSets were split from tests for randomGeneratingSet.