how-to: contributing author
Apparently, this is a proper way to do contributing authors in M2. Not
everyone who wrote some functions is a full “author” of the package.
Authors will be willing to answer questions from M2 users about the
package. Contributing authors need not, but they have made a
significant contribution to the package still.