removed ";" from doc example
Let the user see the ideals, so they can tell if the output makes any
sense at all or not. (with suppressed output, they have to guess what B
is,  making it slightly more difficult to understand the function we
are showcasing.)
also changed N=2 (from 3) since it was giving 3 copies of the same
ideal anyway.